Shadow code merge fix jdbc ut.
