--- shardingsphere-jdbc/shardingsphere-jdbc-core/src/test/resources/config/config-future-shadow.yaml
@@ -1,69 +1,0 @@
-#
-# Licensed to the Apache Software Foundation (ASF) under one or more
-# contributor license agreements.  See the NOTICE file distributed with
-# this work for additional information regarding copyright ownership.
-# The ASF licenses this file to You under the Apache License, Version 2.0
-# (the "License"); you may not use this file except in compliance with
-# the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-schemaName: logic_db
-
-props:
-  sql-show: true
-
-rules:
-- !SHADOW
-  column: shadow
-  sourceDataSourceNames:
-    - shadow_jdbc_0
-  shadowDataSourceNames:
-    - shadow_jdbc_1
-
-  enable: true
-  dataSources:
-    shadowDataSource0:
-      sourceDataSourceName: shadow_jdbc_0
-      shadowDataSourceName: shadow_jdbc_1
-  tables:
-    t_encrypt:
-      dataSourceNames:
-        - shadowDataSource0
-      shadowAlgorithmNames:
-        - id-insert-match-algorithm
-        - id-update-match-algorithm
-        - id-delete-match-algorithm
-        - id-select-match-algorithm
-  shadowAlgorithms:
-    id-insert-match-algorithm:
-      type: COLUMN_REGEX_MATCH
-      props:
-        operation: insert
-        column: id
-        regex: "[1]"
-    id-update-match-algorithm:
-      type: COLUMN_REGEX_MATCH
-      props:
-        operation: update
-        column: id
-        regex: "[1]"
-    id-delete-match-algorithm:
-      type: COLUMN_REGEX_MATCH
-      props:
-        operation: delete
-        column: id
-        regex: "[1]"
-    id-select-match-algorithm:
-      type: COLUMN_REGEX_MATCH
-      props:
-        operation: select
-        column: id
-        regex: "[1]"
